+++ popkorn-example/src/main/kotlin/cc/popkorn/example/model/Ri.kt
@@ -18,3 +18,5 @@ interface R8i
 
 interface R9i
 
+interface R0i
+