+++ popkorn/src/main/kotlin/cc/popkorn/exceptions/AlreadyInjectableException.kt
@@ -1,0 +1,4 @@
+package cc.popkorn.exceptions
+
+class AlreadyInjectableException : RuntimeException("You are trying to add an injectable that is already defined") {
+}