+++ popkorn/src/main/kotlin/cc/popkorn/exceptions/InstanceNotFoundException.kt
@@ -1,0 +1,5 @@
+package cc.popkorn.exceptions
+
+import kotlin.reflect.KClass
+
+class InstanceNotFoundException : RuntimeException("Invalid instance. Seems like you didn't call addInjectable")