+++ popkorn/src/main/kotlin/cc/popkorn/instances/RuntimeInstances.kt
@@ -1,5 +1,7 @@
 package cc.popkorn.instances
 
+import cc.popkorn.exceptions.InstanceNotFoundException
+
 /**
  * Instances implementation for manually scoped instances
  * Calling get() will return a previously added T::class. If not added or already removed,
@@ -12,7 +14,7 @@ internal class RuntimeInstances<T:Any>: Instances<T> {
     private val instances = HashMap<String?, T>()
 
     override fun get(environment:String?) : T{
-        return instances[environment] ?: instances[null] ?: throw RuntimeException("Invalid instance. Seems like you didn't call addInjectable")
+        return instances[environment] ?: instances[null] ?: throw InstanceNotFoundException()
     }
 
     fun put(environment:String?, data:T) = instances.put(environment, data)