+++ popkorn/src/main/kotlin/cc/popkorn/resolvers/RuntimeResolver.kt
@@ -1,6 +1,7 @@
 package cc.popkorn.resolvers
 
 import cc.popkorn.core.Resolver
+import cc.popkorn.exceptions.InstanceNotFoundException
 import kotlin.reflect.KClass
 
 /**
@@ -13,7 +14,7 @@ internal class RuntimeResolver : Resolver<Any> {
     private val resolvers = HashMap<String?, KClass<out Any>>()
 
     override fun resolve(environment: String?): KClass<out Any> {
-        return resolvers[environment] ?: resolvers[null] ?: throw RuntimeException("Invalid instance. Seems like you didn't call addInjectable")
+        return resolvers[environment] ?: resolvers[null] ?: throw InstanceNotFoundException()
     }
 
     fun put(environment:String?, data:KClass<out Any>) = resolvers.put(environment, data)